add warning and move tslib base model to a new file
